+++ client/src/components/common/AuthComponent/index.tsx
