+++ client/src/components/common/Avatar/index.tsx
