+++ client/src/components/common/Post/delete-button.tsx
@@ -2,7 +2,11 @@ import React from 'react';
 import { useRouter } from 'next/router';
 
 import { useModal } from '@/hooks';
-import { ModalType, successMessage, warningMessage } from '@/constant';
+import {
+  ModalType,
+  successMessage,
+  warningMessage,
+} from '@/constants/constant';
 
 //service
 import { deleteCommtent } from '@/service/delete-comment';
