+++ client/src/components/common/Post/post-detail.tsx
