+++ client/src/components/index.tsx
@@ -1,14 +1,19 @@
-export { default as Advertisement } from '@/components/Advertisement';
-export { default as Layout } from '@/components/Layout';
-export { default as MenuView } from '@/components/MenuView';
-export { default as Navigator } from '@/components/Navigator';
-export { Post, PostDetail, Comment, CommentUpload } from '@/components/Post';
-export { default as QuillEditBox } from '@/components/QuillEditBox';
-export { default as Divider } from '@/components/Divider';
-export { default as Avatar } from '@/components/Avatar';
-export { default as LoadingIcon } from '@/components/LoadingIcon';
-export { default as Icon } from '@/components/Icon';
-export { default as Modal } from '@/components/Modal';
-export { default as AuthComponent } from '@/components/AuthComponent';
-export { default as EditProfile } from '@/components/EditProfile';
-export { default as UserInfo } from '@/components/UserInfo';
+export { default as Advertisement } from '@/components/common/Advertisement';
+export { default as Layout } from '@/components/common/Layout';
+export { default as MenuView } from '@/components/common/MenuView';
+export { default as Navigator } from '@/components/common/Navigator';
+export {
+  Post,
+  PostDetail,
+  Comment,
+  CommentUpload,
+} from '@/components/common/Post';
+export { default as QuillEditBox } from '@/components/common/QuillEditBox';
+export { default as Divider } from '@/components/common/Divider';
+export { default as Avatar } from '@/components/common/Avatar';
+export { default as LoadingIcon } from '@/components/common/LoadingIcon';
+export { default as Icon } from '@/components/common/Icon';
+export { default as Modal } from '@/components/common/Modal';
+export { default as AuthComponent } from '@/components/common/AuthComponent';
+export { default as EditProfile } from '@/components/common/EditProfile';
+export { default as UserInfo } from '@/components/common/UserInfo';
